step-3/requirements.txt
attrs==20.3.0
iniconfig==1.1.1
Markdown==3.3.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
py4j==0.10.9
pyparsing==2.4.7
pyspark==3.0.1
pytest==6.2.2
toml==0.10.2

python-dateutil~=2.8.2